Continuing to plug away. Changed the dice from a class to a module with a single function. No use using a class when I'd have to then instantiate a dice to even use one. Now it's as simple as dice.roll().
As for the player, I'm getting it ready for some of the other portions
of the game that we will get to at some point. I have also commented
everything so that it will hopefully be easier to follow if anyone comes
up behind me and wants to add something to it.

Starting to work on the game class itself as well. Even starting the
game is proving tricky as there is a roll that needs to occur which
may or may not require input first by the users.